+++ project-minimization-plugin/src/main/kotlin/org/plan/research/minimization/plugin/psi/stub/KtBlockExpressionStub.kt
@@ -1,4 +1,4 @@
-package org.plan.research.minimization.plugin.model.psi
+package org.plan.research.minimization.plugin.psi.stub
 
 import com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.psi.KtBlockExpression
