+++ project-minimization-plugin/src/main/kotlin/org/plan/research/minimization/plugin/psi/stub/StubDescriptor.kt
@@ -1,4 +1,4 @@
-package org.plan.research.minimization.plugin.model.psi
+package org.plan.research.minimization.plugin.psi.stub
 
 enum class StubDescriptor {
     BLOCK_EXPRESSION, CLASS, CLASS_BODY, FUNCTION, OBJECT, PARAMETER, PROPERTY,
