+++ project-minimization-plugin/src/test/kotlin/hierarchy/file/FileTreeHierarchyGeneratorTest.kt
@@ -1,3 +1,5 @@
+package hierarchy.file
+
 import arrow.core.Either
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.guessProjectDir
