+++ project-minimization-plugin/src/test/kotlin/psi/manager/MinimizationPsiManagerServiceGettingTest.kt
@@ -1,3 +1,5 @@
+package psi.manager
+
 import arrow.core.compareTo
 import com.intellij.openapi.application.readAction
 import com.intellij.openapi.components.service
